course replated CRUD operation

+++ CourseManagementBackend/src/main/java/com/app/dao/CourseDao.java
@@ -4,10 +4,12 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
+import com.app.entity.BatchCycle;
 import com.app.entity.Course;
 
-public interface CourseDao extends JpaRepository<Course, Integer> {
+public interface CourseDao extends JpaRepository<Course, Long> {
 	
 	Optional<Course> findByName(String name);
+
 
 }
